+++ KarhooSDK/Service/Quote/Interactor/VehicleRulesInteractor.swift
@@ -1,0 +1,12 @@
+//
+//  VehicleRulesInteractor.swift
+//  KarhooSDK
+//
+//  Created by Aleksander Wedrychowski on 19/07/2022.
+//  Copyright © 2022 Flit Technologies Ltd. All rights reserved.
+//
+
+import Foundation
+
+protocol VehicleRulesInteractor: KarhooExecutable {
+}
