+++ KarhooSDK/Service/Quote/KarhooQuoteService.swift
@@ -47,48 +47,3 @@ final class KarhooQuoteService: QuoteService {
         return Call(executable: vehicleRulesInteractor)
     }
 }
-
-public struct VehicleRules: Codable, KarhooCodableModel {
-    let rules: [VehicleRule]
-
-    enum CodingKeys: String, CodingKey {
-        case rules = "mappings"
-    }
-}
-
-public struct VehicleRule: Codable {
-    let type: String
-    let tags: [String]
-    let imagePath: String
-    
-    enum CodingKeys: String, CodingKey {
-        case type
-        case tags
-        case imagePath = "image_png"
-    }
-}
-
-protocol VehicleRulesInteractor: KarhooExecutable {
-}
-
-final class KarhooVehicleRulesInteractor: VehicleRulesInteractor {
-    private var requestSender: RequestSender
-
-    init(requestSender: RequestSender = KarhooRequestSender(httpClient: TokenRefreshingHttpClient.shared)) {
-        self.requestSender = requestSender
-    }
-
-    func execute<T>(callback: @escaping CallbackClosure<T>) where T : KarhooCodableModel {
-       
-        requestSender.requestAndDecode(
-            payload: nil,
-            endpoint: .vehicleRules,
-            callback: callback
-        )
-    }
-    
-    func cancel() {
-        requestSender.cancelNetworkRequest()
-    }
-
-}
